# src/main/resources/static/swagger-apis/api1/openapi.yaml
openapi: 3.0.0
info:
  title: Maveric Digital Interview Process APIs
  description: REST APIs supporting Maveric Digital Interview process. Mainly contains APIs for candidate registration, login, taking tests, storing test results, viewing candidate details and test reports
  version: 0.0.1
servers:
  - url: http:/localhost:8080/
    description: Local development server
paths:
...
  /tests:
    get:
      summary: To get the list of all tests.
      description: API to get the list of all tests.
      operationId: getTests
      tags:
        - Tests
      parameters:
        - in: header
          name: Authorization
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: List of all tests
          content:
            application/json:
              schema: 
                type: array
                items: 
                  $ref: '#/components/schemas/Test'
        '401':
          $ref: '#/components/responses/Unauthorized'   
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/UnknownError'
    post:
      summary: To add a new test.
      description: API to register a new test.
      operationId: registerTest
      tags:
        - Tests
      parameters:
        - in: header
          name: Authorization
          required: true
          schema: 
            type: string
      requestBody:
        description: Test details to be added
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Test'
      responses:
        '201':    # status code
          description: Test successfully added
        '401':
          $ref: '#/components/responses/Unauthorized'   
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/UnknownError'
  /tests/{testId}:
    get:
      summary: To get the details of a employee
      description: API to get the details of a employee
      tags:
        - Tests
      operationId: getTestDetails
      parameters:
        - in: header
          name: Authorization
          required: true
          schema: 
            type: string
        - name: testId
          in: path
          description: Test ID
          required: true
          schema:
            type: string
      responses:
        '200':    # status code
          description: Test details successfully obtained
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Candidate'
        '401':
          $ref: '#/components/responses/Unauthorized'   
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/UnknownError'
    put:
      summary: To update a test details
      description: API to update a test details
      tags:
        - Tests
      operationId: updateTest
      parameters:
        - in: header
          name: Authorization
          required: true
          schema: 
            type: string
        - name: testId
          in: path
          description: Test ID
          required: true
          schema:
            type: string
      requestBody:
        description: Test details to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Candidate'
      responses:
        '200':    # status code
          description: Test details successfully updated
        '401':
          $ref: '#/components/responses/Unauthorized'   
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/UnknownError'
    delete:
      summary: To delete a test
      description: API to delete a test 
      tags:
        - Tests
      operationId: deleteTest
      parameters:
        - in: header
          name: Authorization
          required: true
          schema: 
            type: string
        - name: testId
          in: path
          description: Test ID
          required: true
          schema:
            type: string
      responses:
        '200':    # status code
          description: Test successfully deleted
        '401':
          $ref: '#/components/responses/Unauthorized'   
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/UnknownError'
...
  /questionAndAnswers:
    get:
      summary: To get the list of all questions.
      description: API to get the list of all questions.
      operationId: getQuestionAndAnswers
      tags:
        - QuestionAndAnswers
      parameters:
        - in: header
          name: Authorization
          required: true
          schema: 
            type: string
      responses:
        '200':
          description: List of all questions
          content:
            application/json:
              schema: 
                type: array
                items: 
                  $ref: '#/components/schemas/QuestionAndAnswers'
        '401':
          $ref: '#/components/responses/Unauthorized'   
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/UnknownError'
    post:
      summary: To add a new question.
      description: API to add a new question.
      operationId: addQuestionAndAnswer
      tags:
        - QuestionAndAnswers
      parameters:
        - in: header
          name: Authorization
          required: true
          schema: 
            type: string
      requestBody:
        description: Question details to be added
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionAndAnswers'
      responses:
        '201':    # status code
          description: Question and answer successfully added
        '401':
          $ref: '#/components/responses/Unauthorized'   
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/UnknownError'
  /questionAndAnswers/{qaId}:
    get:
      summary: To get the details of a question
      description: API to get the list of all details of a question
      tags:
        - QuestionAndAnswers
      operationId: getQuestionDetails
      parameters:
        - in: header
          name: Authorization
          required: true
          schema: 
            type: string
        - name: qaId
          in: path
          description: Question ID
          required: true
          schema:
            type: string
      responses:
        '200':    # status code
          description: Getting question details was successful
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Candidate'
        '401':
          $ref: '#/components/responses/Unauthorized'   
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/UnknownError'
    put:
      summary: To update a question details
      description: API to update a question details
      tags:
        - QuestionAndAnswers
      operationId: updateQuestionAndAnswers
      parameters:
        - in: header
          name: Authorization
          required: true
          schema: 
            type: string
        - name: qaId
          in: path
          description: Question ID
          required: true
          schema:
            type: string
      requestBody:
        description: Question details to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Candidate'
      responses:
        '200':    # status code
          description: Question details successfully updated
        '401':
          $ref: '#/components/responses/Unauthorized'   
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/UnknownError'
    delete:
      summary: To delete a question
      description: API to delete a question
      tags:
        - QuestionAndAnswers
      operationId: deleteQuestion
      parameters:
        - in: header
          name: Authorization
          required: true
          schema: 
            type: string
        - name: qaId
          in: path
          description: Question ID
          required: true
          schema:
            type: string
      responses:
        '200':    # status code
          description: Question successfully deleted
        '401':
          $ref: '#/components/responses/Unauthorized'   
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/UnknownError'
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    UnknownError:
      description: Unknown error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
    Candidate:
      type: object
      properties:
        _id:
          type: string
          description: The candidate ID.
        firstName:
          type: string
          description: Candidate's first name
        lastName:
          type: string
          description: Candidate's last name
        middleName:
          type: string
          description: Candidate's middle name
        dob:
          type: string
          format: date
          description: Candidate's date of birth
        experience:
          type: number
          description: Candidate's total experience
        skillset:
          type: string
          description: Brief description about Candidate's technical skillset
        panNo:
          type: string
          description: Candidate's PAN number
        email:
          type: string
          description: Candidate's email id
        phoneNo:
          type: number
          description: Candidate's phone number
        noticePeriod:
          type: number
          description: Candidate's notice period in months
        gender:
          type: string
          description: Candidate's gender (Male/Female)
...
    Answer:
      type: object
      properties:
        _id:
          type: string
          description: Unique id to identify this answer
        ans:
          type: string
          description: One of the answer in the answers array
        isCorrectAns:
          type: boolean
          description: Indidates wheather this is the right answer or not
    QuestionAndAnswers:
      type: object
      properties:
        _id: 
          type: string
          description: Unique ID to identify the question
        complexityType:
          type: string
          enum: ["easy", "medium", "hard", "hardest"]
          description: Complexity of the question like easy, medium, hard, hardest
        question:
          type: string
          description: Question to be answered
        answers:
          type: array
          items:
            $ref : "#/components/schemas/Answer"
          description: Array of Answer objects
...
    Test:
      type: object
      properties:
        _id: 
          type: string
          description: Unique ID to identify the test
        subject:
          type: string
          description: Name of the test subject like angular, java etc
        complexityType:
          type: string
          enum: ["easy", "medium", "hard", "hardest"]
          description: Complexity of the test based on experience level
        noOfQuestions:
          type: number
          description: Number of questions in this test
        timeToAnswerInMin:
          type: number
          description: Time available to answer this test in minutes
        cutOffPercentage:
          type: number
          description: Cut off percentage 
...